pyproject.toml
[project]
name = "transcribe-with-whisper"
version = "0.8.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asteroid-filterbanks>=0.4.0",
    "fastapi==0.118.0",
    "faster-whisper==1.2.0",
    "htmldocx>=0.0.6",
    "huggingface-hub==0.35.3",
    "pyannote-audio==4.0.0",
    "pydub==0.25.1",
    "python-docx>=0.8.12",
    "python-multipart==0.0.20",
    "torch==2.8.0",
    "uvicorn[standard]==0.37.0",
    "webvtt-py==0.5.1",
]
